removed a lot of warnings

+++ src/LayerActivation.cpp
@@ -18,7 +18,7 @@ void LayerActivation::forward(const MatrixFloat& mIn,MatrixFloat& mOut) const
 {
     mOut.resize(mIn.rows(),mIn.cols());
 
-    for(size_t i=0;i<mOut.size();i++)
+    for(int i=0;i<mOut.size();i++)
         mOut(i)=_pActivation->apply(mIn(i));
 }
 ///////////////////////////////////////////////////////////////////////////////
@@ -27,7 +27,7 @@ void LayerActivation::backpropagation(const MatrixFloat &mInput,const MatrixFloa
     (void)fLearningRate;
 
     mNewDelta.resize(mDelta.rows(),mDelta.cols());
-    for(size_t i=0;i<mNewDelta.size();i++)
+    for(int i=0;i<mNewDelta.size();i++)
         mNewDelta(i)=_pActivation->derivation(mInput(i))*mDelta(i);
 }
 ///////////////////////////////////////////////////////////////////////////////
