+++ src/Matrix.cpp
@@ -1,6 +1,6 @@
 #include "Matrix.h"
 
-const MatrixFloat from_raw_buffer(float *pBuffer,size_t iRows,size_t iCols)
+const MatrixFloat from_raw_buffer(float *pBuffer,int iRows,int iCols)
 {
 #ifdef USE_EIGEN
     return Eigen::Map<MatrixFloat>((float*)pBuffer,static_cast<Eigen::Index>(iRows),static_cast<Eigen::Index>(iCols));
