+++ src/NetTrainLearningRate.cpp
@@ -14,7 +14,7 @@ NetTrainLearningRate::~NetTrainLearningRate()
 /////////////////////////////////////////////////////////////////////////////////////////////////
 void NetTrainLearningRate::train(Net& net,const MatrixFloat& mSamples,const MatrixFloat& mTruth,const TrainOption& topt)
 {
-    size_t nLayers=net.layers().size();
+    int nLayers=(int)net.layers().size();
 
     if(nLayers==0)
     {
@@ -25,22 +25,22 @@ void NetTrainLearningRate::train(Net& net,const MatrixFloat& mSamples,const Matr
 
     if(topt.initWeight)
     {
-        for(unsigned int i=0;i<nLayers;i++)
+        for(int i=0;i<nLayers;i++)
             net.layer(i)->init();
     }
 
     //TrainResult tr;
-    size_t iBatchSize=topt.batchSize;
-    size_t iNbSamples=mSamples.rows();
+    int iBatchSize=topt.batchSize;
+    int iNbSamples=mSamples.rows();
 
     for(int iEpoch=0;iEpoch<topt.epochs;iEpoch++)
     {
         MatrixFloat mShuffle=rand_perm(iNbSamples);
 
-        size_t iBatchStart=0;
+        int iBatchStart=0;
         while(iBatchStart<iBatchSize)
         {
-            size_t iBatchEnd=iBatchStart+iBatchSize;
+            int iBatchEnd=iBatchStart+iBatchSize;
             if(iBatchEnd>iBatchSize) // do not compute with partial minibatch ? todo?
                 break;
 
@@ -51,18 +51,18 @@ void NetTrainLearningRate::train(Net& net,const MatrixFloat& mSamples,const Matr
             MatrixFloat sumDelta; //todo keep object memory beetween minbatch, avoid malloc
 
             //forward pass, and compute mean layer input and error
-            for(size_t iSample=iBatchStart;iSample<iBatchEnd;iSample++)
+            for(int iSample=iBatchStart;iSample<iBatchEnd;iSample++)
             {
                 //compute total error, sample by sample
 
                 //forward pass with layer input save
-                size_t iIndexSample=(size_t)mShuffle(iSample,0);
+                int iIndexSample=(int)mShuffle(iSample,0);
                 const MatrixFloat& mSample=mSamples.row(iIndexSample);
                 const MatrixFloat& mTarget=mTruth.row(iIndexSample);
                 MatrixFloat mOut, mTemp=mSample;
 
                 //forward pass with store
-                for(unsigned int i=0;i<nLayers;i++)
+                for(int i=0;i<nLayers;i++)
                 {
                     net.layer(i)->forward(mTemp,mOut);
                     MatrixFloat& mIn=sumInput[i];
@@ -83,7 +83,7 @@ void NetTrainLearningRate::train(Net& net,const MatrixFloat& mSamples,const Matr
             }
 
             // normalize vs batchsize
-            for(unsigned int i=0;i<nLayers;i++)
+            for(int i=0;i<nLayers;i++)
                 sumInput[i]/=(float)iBatchSize;
 
             //backpropagation of delta and update of weights
